Fix install as well
